+++ src/components/pages/home/index.ts
@@ -1,8 +1,5 @@
 export { HeroSection } from "./HeroSection";
 export { TrustStrip } from "./TrustStrip";
 export { PackageCardsSection } from "./PackageCardsSection";
-export { HowItWorksSection } from "./HowItWorksSection";
-export { CoachingFocusSection } from "./CoachingFocusSection";
-export { TestimonialsSection } from "./TestimonialsSection";
 export { FAQPreviewSection } from "./FAQPreviewSection";
 export { CTABandSection } from "./CTABandSection";